Dynamic Font loading fully functional for 1 runtime generated font

+++ Harfbuzz/SDF/SignedDistanceFixed.cs
@@ -6,6 +6,6 @@ namespace HarfBuzz.SDF
     {
         public int distance;    //FT_16D16
         public int cross;       //FT_16D16
-        public sbyte sign;
+        public int sign;
     }
 }
